+++ test/unit/device-api/actions.spec.ts
@@ -1,10 +1,9 @@
 import { expect } from 'chai';
 import type { SinonStub } from 'sinon';
-import { spy, useFakeTimers, stub } from 'sinon';
+import { stub } from 'sinon';
 
 import * as hostConfig from '~/src/host-config';
 import * as actions from '~/src/device-api/actions';
-import blink = require('~/lib/blink');
 
 describe('device-api/actions', () => {
 	describe('runs healthchecks', () => {
@@ -24,25 +23,6 @@ describe('device-api/actions', () => {
 			expect(await actions.runHealthchecks([taskFalse, taskError])).to.be.false;
 			expect(await actions.runHealthchecks([taskFalse, taskFalse])).to.be.false;
 			expect(await actions.runHealthchecks([taskError, taskError])).to.be.false;
-		});
-	});
-
-	describe('identifies device', () => {
-		// This suite doesn't test that the blink submodule writes to the correct
-		// led file location on host. That test should be part of the blink module.
-		it('directs device to blink for set duration', async () => {
-			const blinkStartSpy = spy(blink.pattern, 'start');
-			const blinkStopSpy = spy(blink.pattern, 'stop');
-			const clock = useFakeTimers();
-
-			actions.identify();
-			expect(blinkStartSpy.callCount).to.equal(1);
-			clock.tick(15000);
-			expect(blinkStopSpy.callCount).to.equal(1);
-
-			blinkStartSpy.restore();
-			blinkStopSpy.restore();
-			clock.restore();
 		});
 	});
